Design+Implementatin
I have used the partially Build Api Project and used Collection View to Display Poke Gallery. In Controller we define the Collection View with extension of Controller.Which have Collection view properties like count total count and define the data source or data at each row. and UICollectionViewDelegateFlowLayout contain the size of cell and the row and column spaces